+++ src/smc_lammps/reader/parser.py
@@ -1,8 +1,11 @@
 from __future__ import annotations
 
+import typing
 from io import StringIO
-from itertools import islice
 from pathlib import Path
+
+if typing.TYPE_CHECKING:
+    from _typeshed import OpenTextMode
 
 import numpy as np
 
@@ -17,8 +20,8 @@ class Parser:
     class EndOfLammpsFile(Exception):
         pass
 
-    def __init__(self, file: Path, time_it: bool = False) -> None:
-        self.file = open(file, "r", encoding="utf-8")
+    def __init__(self, file: Path, time_it: bool = False, mode: OpenTextMode = "r") -> None:
+        self.file = open(file, mode, encoding="utf-8")
 
         self.timings = None
         if time_it:
@@ -31,7 +34,9 @@ class Parser:
         current_line = None
         empty = True
 
-        for line in self.file:
+        # NOTE: use readline instead of a for loop,
+        # since the latter breaks file.seek() calls
+        while line := self.file.readline():
             empty = False
 
             if line.startswith("ITEM: ATOMS"):
@@ -76,9 +81,14 @@ class Parser:
         timestep = int(saved["ITEM: TIMESTEP"][0])
         number_of_atoms = int(saved["ITEM: NUMBER OF ATOMS"][0])
 
-        lines = list(islice(self.file, number_of_atoms))
-        if len(lines) != number_of_atoms:
-            raise ValueError("reached end of file unexpectedly")
+        lines: list[str] = []
+        for _ in range(number_of_atoms):
+            try:
+                next = self.file.readline()
+            except Exception as e:
+                raise ValueError("reached end of file unexpectedly") from e
+            else:
+                lines.append(next)
 
         data = self.split_data(self.get_array(lines))